Add hack to see if pre-encoded args will work
