+++ services/construction/payloads.go
@@ -100,7 +100,7 @@ func (s *APIService) ConstructionPayloads(
 			return nil, sdkTypes.WrapErr(sdkTypes.ErrInvalidInput, err)
 		}
 
-		data, err := constructContractCallData(metadata.MethodSignature, metadata.MethodArgs)
+		data, err := constructContractCallDataGeneric(metadata.MethodSignature, metadata.MethodArgs)
 		if err != nil {
 			return nil, sdkTypes.WrapErr(sdkTypes.ErrInvalidInput, err)
 		}
